Extract `AbstractFluxConcatMapTest`
